break: rewrite compile includes prod mode

+++ src/compile/index.ts
@@ -30,21 +30,25 @@ const compileJSX = (
   }
 };
 
-const copyRestFilesToTsc = (input: string, outputPrefix: string) =>
-  walk(input)
-    .filter(file => {
+const copyRestFilesToTsc = async (input: string, outputPrefix: string) => {
+  const files: any = await walk(input);
+  files
+    .filter((file: any) => {
       const ext = path.extname(file);
       if (ext !== '.ts' && ext !== '.tsx') {
         return true;
       }
     })
-    .filter(file => path.basename(file) !== 'tsconfig.json')
-    .map(file => {
-      fs.copySync(
-        joinWithRootPath(file),
-        joinWithRootPath(file.replace(input, outputPrefix)),
-      );
+    .filter((file: any) => path.basename(file) !== 'tsconfig.json')
+    .map((file: any) => {
+      if (file !== file.replace(input, outputPrefix)) {
+        fs.copySync(
+          joinWithRootPath(file),
+          joinWithRootPath(file.replace(input, outputPrefix)),
+        );
+      }
     });
+};
 
 // if only tsx, compile them to jsx with tsc
 // then compile them to es5 with babel
@@ -53,31 +57,44 @@ export const compile = async (program: any, callback?: () => void) => {
   const entry = program.entry || 'src';
   const output = program.output || 'lib';
   const outputEs = program.outputEs || 'es';
+  const isTestEnv = process.env.RUN_ENV === 'test';
 
-  if (process.env.RUN_ENV === 'test') {
-    const tscOutput = 'dist/test-cases';
-    const cleanPaths = [output, outputEs, tscOutput];
-    for (const item of cleanPaths) {
-      const target = joinWithRootPath(item);
-      if (fs.existsSync(target)) {
-        fs.emptyDirSync(target);
-      }
+  let tscOutputPath = 'dist';
+  if (fs.existsSync(joinWithRootPath('tsconfig.json')) && !tscOutputPath) {
+    const tsconfigFile = require(joinWithRootPath('tsconfig.json'));
+    tscOutputPath = tsconfigFile.compilerOptions.outDir;
+  }
+
+  const tscOutput = isTestEnv ? 'dist/test-cases' : tscOutputPath;
+  const cleanPaths = [output, outputEs, tscOutput];
+  for (const item of cleanPaths) {
+    const target = joinWithRootPath(item);
+    if (fs.existsSync(target)) {
+      fs.emptyDirSync(target);
     }
-    trace(`少女边清理着名为 ${output} 的钱箱，边回顾着即将结束的一年单身生活
-...顺带感慨了下自己又一年一平如洗的身板`);
+  }
+  trace(
+    `少女边清理着名为 ${output}/${outputEs} 的钱箱，边回顾着即将结束的一年单身生活`,
+  );
+  setTimeout(() => {
+    trace(`...顺带感慨了下自己又一年一平如洗的身板`);
+  }, 2000);
 
+  await new Promise(resolve => {
     const tscBin = require.resolve('typescript/bin/tsc');
-    await new Promise(resolve => {
+    if (isTestEnv) {
       runCmd('node', [tscBin, '-p', entry], resolve);
-    });
-    copyRestFilesToTsc(entry, tscOutput);
-
-    const es6Files = walk(joinWithRootPath(tscOutput));
-    compileJSX(es6Files, tscOutput, output, 'es2015');
-    compileJSX(es6Files, tscOutput, outputEs, 'es2015+');
-    info('少女换上了新的钱箱，开始了一年新的单身生活');
-    if (callback) {
-      callback();
+    } else {
+      runCmd('node', [tscBin], resolve);
     }
+  });
+
+  await copyRestFilesToTsc(entry, tscOutput);
+  const tscOutputFiles: any = await walk(joinWithRootPath(tscOutput));
+  compileJSX(tscOutputFiles, tscOutput, output, 'es2015');
+  compileJSX(tscOutputFiles, tscOutput, outputEs, 'es2015+');
+  info('少女换上了新的钱箱，开始了一年新的单身生活');
+  if (callback) {
+    callback();
   }
 };
